+++ Calc/Model/ResultState.cs
@@ -37,14 +37,14 @@ namespace Calc.Model
             calculatorModel.state = State.Opers;
             calculatorModel.MathOperator = buttonContent;
             calculatorModel.CanBeRefreshed = true;
-            calculatorModel.SetFirstNumber(double.Parse(calculatorModel.DialText));
+            calculatorModel.SetFirstNumber(decimal.Parse(calculatorModel.DialText));
             string v = calculatorModel.FirstNumber + calculatorModel.MathOperator;
             calculatorModel.SetLastOperation(v);
         }
 
         public void OnResultClicked(CalculatorModel calculatorModel)
         {
-            calculatorModel.SetFirstNumber(double.Parse(calculatorModel.DialText));
+            calculatorModel.SetFirstNumber(decimal.Parse(calculatorModel.DialText));
             calculatorModel.Calculate();
         }
     }
